[GPS] Improved precision
Applied PR from @qbazd
[lat, lng force math operations #47](https://github.com/mikalhart/TinyGPSPlus/pull/47)

+++ lib/TinyGPSPlus-1.0.2/src/TinyGPS++.cpp
@@ -355,14 +355,14 @@ void TinyGPSLocation::setLongitude(const char *term)
 double TinyGPSLocation::lat()
 {
    updated = false;
-   double ret = rawLatData.deg + rawLatData.billionths / 1000000000.0;
+   double ret = (double)rawLatData.deg + ((double)rawLatData.billionths / (double)1000000000.0);
    return rawLatData.negative ? -ret : ret;
 }
 
 double TinyGPSLocation::lng()
 {
    updated = false;
-   double ret = rawLngData.deg + rawLngData.billionths / 1000000000.0;
+   double ret = (double)rawLngData.deg + ((double)rawLngData.billionths / (double)1000000000.0);
    return rawLngData.negative ? -ret : ret;
 }
